# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: trb-core
  description: Ядро trust-bank
  version: "1"
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/transactions/withdrawal:
    post:
      tags:
        - API для счетов
      summary: Вывести деньги
      operationId: withdrawal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnidirectionalTransactionDto'
        required: true
      responses:
        "200":
          description: OK
  /api/v1/transactions/replenishment:
    post:
      tags:
        - API для счетов
      summary: Пополнить счет
      operationId: replenishment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnidirectionalTransactionDto'
        required: true
      responses:
        "200":
          description: OK
  /api/v1/accounts:
    post:
      tags:
        - API для счетов
      summary: Создать счет для клиента
      operationId: createClientAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAccountDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDto'
  /api/v1/users/{id}/accounts:
    get:
      tags:
        - API для счетов
      summary: Получить счета клиента
      operationId: getUserAccounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDto'
  /api/v1/accounts/{id}:
    get:
      tags:
        - API для счетов
      summary: Получить информацию о счете по id
      operationId: getAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountDto'
    delete:
      tags:
        - API для счетов
      summary: Закрыть счет
      operationId: closeAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
  /api/v1/accounts/{accountId}/history:
    get:
      tags:
        - API для счетов
      summary: Получить историю по счету
      operationId: getHistory
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: true
          schema:
            minimum: 0
            type: integer
            format: int32
        - name: size
          in: query
          required: true
          schema:
            maximum: 200
            minimum: 1
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseTransactionDto'
components:
  schemas:
    UnidirectionalTransactionDto:
      required:
        - accountId
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        amount:
          minimum: 1
          type: integer
          format: int64
    NewAccountDto:
      required:
        - clientFullName
        - externalClientId
        - type
      type: object
      properties:
        type:
          type: string
          description: Тип счета
          example: DEPOSIT
          enum:
            - DEPOSIT
            - LOAN
            - MASTER
        clientFullName:
          type: string
          description: ФИО владельца счета
          example: Иванов Иван Иванович
        externalClientId:
          type: string
          description: Идентификатор пользователя из trb-users
          format: uuid
    AccountDto:
      required:
        - balance
        - clientFullName
        - creationDate
        - externalClientId
        - id
        - isClosed
        - type
      type: object
      properties:
        id:
          type: string
          description: Идентификатор счета
          format: uuid
        type:
          type: string
          description: Тип банковского счета
          enum:
            - DEPOSIT
            - LOAN
            - MASTER
        balance:
          type: integer
          description: Баланс счета
          format: int64
        clientFullName:
          type: string
          description: ФИО владельца счета
        externalClientId:
          type: string
          description: Внешний идентификатор пользователя из trb-users
        creationDate:
          type: string
          description: Дата создания счета
          example: "1709372199882"
        closingDate:
          type: string
          description: Дата закрытия счета
          example: "1709372199882"
        isClosed:
          type: boolean
          description: Признак закрытия счета
    PaginationResponseTransactionDto:
      required:
        - elements
        - pageNumber
        - pageSize
      type: object
      properties:
        pageNumber:
          type: integer
          description: Номер страницы
          format: int32
        pageSize:
          type: integer
          description: Размер страницы
          format: int32
        elements:
          type: array
          description: Элементы
          items:
            $ref: '#/components/schemas/TransactionDto'
    TransactionDto:
      required:
        - amount
        - code
        - date
        - id
        - state
        - type
      type: object
      properties:
        id:
          type: string
          description: Идентификатор транзакции
          format: uuid
        date:
          type: string
          description: Дата проведения транзакции
          example: "1709372199882"
        payerAccountId:
          type: string
          description: Отправитель
          format: uuid
        payeeAccountId:
          type: string
          description: Получатель
          format: uuid
        amount:
          type: integer
          description: Сумма
          format: int64
        type:
          type: string
          description: Тип транзакции
          enum:
            - TRANSFER
            - REPLENISHMENT
            - WITHDRAWAL
            - LOAN_REPAYMENT
            - LOAN_PAYMENT
        state:
          type: string
          description: Состояние транзакции
          enum:
            - DONE
            - REJECTED
        code:
          type: string
          description: Код транзакции
          enum:
            - SUCCESS
            - NOT_ENOUGH_MONEY
      description: Элементы